fix: mini refactoring
