+++ src/logic/animation.ts
@@ -53,4 +53,3 @@ export const animationLogic = () => {
     showAnimation();
   }, 1500);
 };
-// animationLogic();
